refactored the code according to the review feedback
